Move confirmation prompt to command level
