+++ packages/office-addin-project/src/commands.ts
@@ -2,6 +2,7 @@
 // Licensed under the MIT license.
 
 import * as commander from "commander";
+import * as inquirer from "inquirer";
 import { logErrorMessage } from "office-addin-usage-data";
 import { usageDataObject } from "./defaults";
 import { convertProject } from "./convert";
@@ -11,10 +12,27 @@ export async function convert(command: commander.Command) {
     const manifestPath: string = command.manifest ?? "./manifest.xml";
     const backupPath: string = command.backup ?? "./backup.zip";
 
-    await convertProject(manifestPath, backupPath);
-    usageDataObject.reportSuccess("convert");
+    const shouldContinue = await asksForUserConfirmation();
+    if (shouldContinue) {
+      await convertProject(manifestPath, backupPath);
+      usageDataObject.reportSuccess("convert", { result: "Project converted" });
+    } else {
+      usageDataObject.reportSuccess("convert", {
+        result: "Conversion cancelled",
+      });
+    }
   } catch (err: any) {
     usageDataObject.reportException("convert", err);
     logErrorMessage(err);
   }
+}
+
+async function asksForUserConfirmation(): Promise<boolean> {
+  const question = {
+    message: `This command will convert your current xml manifest to a json manifest and then proceed to upgrade your project dependencies to ensure compatibility with the new project structure.\nHowever, in order for this newly updated project to function correctly you must be on a private environment that has not yet been released publicly.\nWould you like to continue?`,
+    name: "didUserConfirm",
+    type: "confirm",
+  };
+  const answers = await inquirer.prompt([question]);
+  return (answers as any).didUserConfirm;
 }
